requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
requests
torch
torchvision
Pillow
ipython